turns out we just need more diffusion steps
